// Home/transaction/fin-11-invoice/model/fin-11.ts
export class Fin11 {
    stateName;
    DistrictName;
    clinicType;
    clinic;
    stateId: number;
    districtId: number;
    clinicTypeId: number;
    clinicCode: string;
    clinicId: number;
    month: string;
    year: string;
    assetName: string;
    typeOfExpenses;
    concessionElements: number;
    reference;
    status: string;
    
}